# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"HistoryParams": {
				"properties": {
					"address": {
						"type": "string"
					},
					"token": {
						"properties": {
							"decimal": {
								"type": "string"
							},
							"chain": {
								"type": "string"
							},
							"address": {
								"type": "string"
							}
						},
						"required": [
							"decimal",
							"chain",
							"address"
						],
						"type": "object"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"address",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "spw_market",
		"version": "2.0.0",
		"description": "build: Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Nighty Eight"
		}
	},
	"paths": {
		"/wallet/approval": {
			"get": {
				"operationId": "GetApproval",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chain",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/wallet/history": {
			"post": {
				"operationId": "GetHistory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/HistoryParams"
							}
						}
					}
				}
			}
		},
		"/log": {
			"post": {
				"operationId": "Record",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/log/dapps": {
			"post": {
				"operationId": "RecordDapp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}